Added method to get coordinates
Can now fetch coordinates of the device.
Added permission check to MapsActivity, and checks these before a picture can be taken.

+++ app/src/main/java/dk/easv/geoshare/BE/PhotoMetaData.java
@@ -8,15 +8,15 @@ import java.sql.Timestamp;
 public class PhotoMetaData implements Serializable {
 
     private String photoID;
-    private Long lat;
-    private Long lng;
+    private double lat;
+    private double lng;
     private Long timestamp;
 
     public PhotoMetaData() {
 
     }
 
-    public PhotoMetaData(Long lat, Long lng, Long timestamp) {
+    public PhotoMetaData(double lat, double lng, Long timestamp) {
         this.lat = lat;
         this.lng = lng;
         this.timestamp = timestamp;
@@ -30,11 +30,11 @@ public class PhotoMetaData implements Serializable {
         return photoID;
     }
 
-    public Long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public Long getLng() {
+    public double getLng() {
         return lng;
     }
